Add "Tables" navigation item to dashboard layout
A new list item (<li>) with a link (<a>) to the "Index" action of the "Table" controller has been added to the `_DashboardLayout.cshtml` file. The link uses the `@IsActive` method to apply the "active" class when the current page is the "Table" page, enabling users to navigate to the "Tables" section of the application.