Add deleteAction function to PayrollView.cshtml
Introduced a new JavaScript function `deleteAction` in the `PayrollView.cshtml` file. This function constructs a URL for the `Delete` action in the `Payroll` controller using a provided `id`, sends a GET request to this URL, and updates the `modalContent` element with the response. It then displays the modal with the ID `modal`, enhancing the user interface by allowing in-page deletion confirmations.